parsewav.py: Tool to measure/verify my figures.
A .WAV parsing tool to take recordings of cassette saves as input,
and provide some information about the recording. I wanted to
confirm my numbers about how many zeroes in a gap, and the rough
frequencies for signal cycles. I was correct on signal frequency
(at least for "zero" cycles), but I was wrong about the gap signal:
I'd thought the gap was 20,000 zeroes, but it's 19,744, which is
256 fewer. Presumably the authors made the same calculation error
that I had, and had actually intended 20,000 zeroes.